test: Refactor into separate spec files.
